split to files and add cypress
