+++ package.json
@@ -5,7 +5,9 @@
   "main": "index.js",
   "scripts": {
     "start": "webpack-serve ./webpack.config.js",
-    "build": "webpack --config webpack.config.prod.js"
+    "build": "webpack --config webpack.config.prod.js",
+    "test": "npx cypress run",
+    "lint:fix": "eslint --fix './src/**/*.{ts,tsx}'"
   },
   "repository": {
     "type": "git",
@@ -39,6 +41,7 @@
     "@types/styled-components": "^4.4.2",
     "@typescript-eslint/eslint-plugin": "^2.15.0",
     "@typescript-eslint/parser": "^2.15.0",
+    "cypress": "^3.8.2",
     "eslint": "^6.8.0",
     "eslint-plugin-react": "^7.17.0",
     "file-loader": "^5.0.2",
